refactor: extract shared text-node transform into utils.ts
The separator-marking technique (mark → concatenate → transform →
split → write back) was duplicated verbatim between remark.ts and
rehype.ts. Extract it into `transformTextNodes()` in utils.ts.

Both plugins now call the shared helper. The rehype `transformElement`
retains its `checkInvariance` wrapper around the shared call.

Also add test and lint gates to the pre-push hook — it previously
built and published to npm without running the test suite, meaning
a broken commit could be published.
